Split http logger, add request logs to body parser
