+++ src/common/http-logger/logger.ts
@@ -1,0 +1,3 @@
+import logger from '../logger'
+
+export default logger.child({ type: 'http' })
